+++ scripts/lichessTools-init.js
@@ -114,6 +114,7 @@ lichessTools.loadTool(LiChessTools.Tools.PuzzleDownload);
 lichessTools.loadTool(LiChessTools.Tools.GameListOptions);
 lichessTools.loadTool(LiChessTools.Tools.ShowDeviation);
 lichessTools.loadTool(LiChessTools.Tools.LinkToAnalysis);
+lichessTools.loadTool(LiChessTools.Tools.PuzzleOptions);
 
 lichessTools.loadTool(LiChessTools.Tools.HideBoardCommand);
 lichessTools.loadTool(LiChessTools.Tools.TrapValueCommand);